refactor: split training-load.ts by responsibility (#585)
Pure refactor — no logic changes. The 994-line `services/training-load.ts`
is replaced by a `services/training-load/` directory split by responsibility:

- banister.ts (400)    — Pure Banister math: TRIMP, hourly impulses, EMA,
                         recovery zones, settings/HR helpers. No DB access.
- query.ts (235)       — Read path: computeTrainingLoad
- recompute.ts (157)   — Write path: chunked impulse recomputation
- deps.ts (123)        — TrainingLoadDeps interface + production deps factory
- aggregation.ts (89)  — DST-aware bucket aggregation (Temporal-based)
- hr-cache.ts (39)     — observed_hr_max cache shared between read + write
- index.ts (36)        — Barrel re-export

Test file moved into the directory and updated to import from `./index.ts`.
Consumer imports in `mcp/training-load-tools.ts` and
`routes/training-load-router.ts` updated to `services/training-load/index.ts`.
